feat: 增加 DockComponentBase 基类
